moved handleSetCurrentCohortChanges into cohortDetails
